+++ src/bzfs/bzfsAPI.cxx
@@ -1170,7 +1170,7 @@ BZF_API bool bz_isPlayerPaused( int playerID )
   return player->player.isPaused();
 }
 
-BZF_API int bz_getPausedTime( int playerID )
+BZF_API double bz_getPausedTime( int playerID )
 {
   GameKeeper::Player *player = GameKeeper::Player::getPlayerByIndex(playerID);
 
@@ -1180,7 +1180,7 @@ BZF_API int bz_getPausedTime( int playerID )
   return player->player.getPausedTime();
 }
 
-BZF_API int bz_getIdleTime( int playerID )
+BZF_API double bz_getIdleTime( int playerID )
 {
   GameKeeper::Player *otherData = GameKeeper::Player::getPlayerByIndex(playerID);
